# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Package Registry Scraper API",
    "description": "Scrapes and tracks metadata from supported package registries like crates.io and jsr.io. Provides asynchronous job submission and tracking.",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "Jobs",
      "description": "Operations related to scraping jobs"
    }
  ],
  "paths": {
    "/jobs": {
      "get": {
        "summary": "List jobs",
        "description": "Retrieves a list of previously submitted scraping jobs. Supports cursor-based pagination.",
        "tags": ["Jobs"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of jobs to return",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Order the jobs by",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"]
            }
          },
          {
            "name": "after",
            "in": "query",
            "description": "Cursor for pagination to fetch jobs after a specific job ID",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jobs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create job",
        "description": "Creates a new scraping job for a specific package registry and package name.",
        "tags": ["Jobs"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequest"
              },
              "example": {
                "registry_name": "crates.io",
                "package_name": "tokio"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/JobResponseWrapper"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{id}": {
      "get": {
        "summary": "Get job by ID",
        "description": "Retrieves a specific scraping job by its unique identifier.",
        "tags": ["Jobs"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the job to retrieve",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponseWrapper"
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateJobRequest": {
        "type": "object",
        "required": ["registry_name", "package_name"],
        "properties": {
          "registry_name": {
            "type": "string",
            "enum": ["crates.io", "jsr.io"],
            "description": "The registry where the package is hosted"
          },
          "package_name": {
            "type": "string",
            "description": "Name of the package to scrape"
          }
        }
      },
      "Job": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier of the job"
          },
          "registry_name": {
            "type": "string",
            "enum": ["crates.io", "jsr.io"],
            "description": "The registry name"
          },
          "package_name": {
            "type": "string",
            "description": "Name of the package"
          },
          "status": {
            "type": "string",
            "enum": ["processing", "complete"],
            "description": "Current status of the job"
          },
          "trace_id": {
            "type": "string",
            "description": "Trace ID for correlating logs"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the job was created"
          }
        },
        "example": {
          "id": "a79f1b0c-85f3-4e45-b3b2-cff1b162a021",
          "registry_name": "crates.io",
          "package_name": "tokio",
          "status": "processing",
          "trace_id": "abc123",
          "created_at": "2025-05-26T14:00:00Z"
        }
      },
      "JobResponseWrapper": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Job"
          }
        }
      },
      "JobListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Job"
            }
          },
          "next_cursor": {
            "type": ["string", "null"],
            "description": "Cursor for fetching the next page of results"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "description": "Describes the nature of the error"
              }
            }
          }
        },
        "example": {
          "error": {
            "message": "Job not found"
          }
        }
      }
    }
  }
}